pyproject.toml
[tool.poetry]
name = "hevy-reporting"
version = "0.1.0"
description = ""
authors = ["Bosire Nyakundi <Nyakundi@hotmail.co.uk>"]
readme = "README.md"
packages = [{include = "hevy"},{include = "reporting"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
requests = "^2.32.3"
asyncio = "^3.4.3"
aiohttp = "^3.10.0"
python-dotenv = "^1.0.1"
async-lru = "^2.0.4"
uvicorn = "^0.30.5"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
